Refactor to use primary constructors and clean up codebase Refactored service, repository, and controller classes to use C# primary constructor syntax, removing redundant field declarations and constructors. Cleaned up unnecessary usings and comments, improved parameter naming consistency, and added XML docs to CountryController endpoints. Updated dependency injection registrations and reorganized stored procedure constants for clarity. Applied formatting improvements throughout for better readability.

+++ HotelAPI/Controller/GlobalSearchController.cs
@@ -1,5 +1,4 @@
 ﻿using HotelAPI.BAL.Interfaces;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace HotelAPI.Controller
@@ -8,16 +7,16 @@ namespace HotelAPI.Controller
 	[Route("api/[controller]")]
 	public class GlobalSearchController(IGlobalSearchService globalSearchService) : ControllerBase
 	{
-
 		/// <summary>
 		/// Global  search (City, Region, District, Hotel, etc.)
 		/// </summary>
 		/// <param name="q">Search text (minimum 2 characters)</param>
+
 		[HttpGet("Globalsearch")]
-		public async Task<IActionResult>Search([FromQuery] string q)
+		public async Task<IActionResult> Search([FromQuery] string q)
 		{
 			var result = await globalSearchService.SearchAsync(q);
-			return StatusCode((int)result.Code, result);
+			return StatusCode(result.Code, result);
 		}
 	}
 }
